requirements.txt
piexif
pillow
pyqt5
sklearn
click
dlib